+++ themegen/lib/src/core/logic/code_producer.dart
@@ -1,0 +1,11 @@
+import 'package:code_builder/code_builder.dart';
+
+abstract class CodeProducer<T> {
+  CodeProducer(this._emitter);
+
+  final DartEmitter _emitter;
+
+  Spec spec(T input);
+
+  String produce(T input) => spec(input).accept(_emitter).toString();
+}