+++ themegen/lib/themegen.dart
@@ -2,17 +2,20 @@ library themegen;
 
 import 'package:build/build.dart' as build;
 import 'package:code_builder/code_builder.dart';
-import 'package:source_gen/source_gen.dart';
-import 'package:themegen/src/theme/generator/theme_generator.dart';
-import 'package:themegen/src/theme/logic/theme_producer.dart';
+import 'package:source_gen/source_gen.dart' as source_gen;
+import 'package:themegen/src/feature/ext/logic/ext_producer.dart';
+import 'package:themegen/src/feature/theme/theme_generator.dart';
+import 'package:themegen/src/feature/theme/theme_producer.dart';
 
 build.Builder themeGen(build.BuilderOptions options) {
   final emitter = DartEmitter(useNullSafetySyntax: true);
-  return SharedPartBuilder(
+  return source_gen.SharedPartBuilder(
     [
       ThemeGenerator(
-        producer: ThemeProducer(emitter),
-        allowedEntity: 'Class',
+        ThemeProducer(
+          emitter,
+          ExtProducer(emitter),
+        ),
       ),
     ],
     'theme',