+++ themegen_annotation/CHANGELOG.md
@@ -11,4 +11,5 @@
 - Added LICENSE
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Added the first version of the themegen_annotation package.
+- Added extensions